+++ task-17/src/main/java/by/slava_borisov/hoteladmin/dao/RoomDao.java
@@ -3,6 +3,7 @@ package by.slava_borisov.hoteladmin.dao;
 import by.slava_borisov.hoteladmin.model.Room;
 import by.slava_borisov.hoteladmin.model.RoomStatus;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -15,7 +16,7 @@ public interface RoomDao extends GenericDao<Room, Long> {
 
     void updateStatus(Long roomId, RoomStatus status);
 
-    void updatePricePerNight(Long roomId, double newPrice);
+    void updatePricePerNight(Long roomId, BigDecimal newPrice);
 
     int countAvailable();
