+++ task-17/src/main/java/by/slava_borisov/hoteladmin/dto/RoomDto.java
@@ -1,9 +1,11 @@
 package by.slava_borisov.hoteladmin.dto;
 
+import java.math.BigDecimal;
+
 public record RoomDto(
         Long id,
         Integer number,
-        double pricePerNight,
+        BigDecimal pricePerNight,
         RoomStatusDto status,
         int capacity,
         int stars
